TNS queue for submission
