+++ config.yaml.defaults
@@ -643,6 +643,7 @@ ports:
   facility_queue: 64510
   notification_queue: 64610
   observation_plan_queue: 64710
+  tns_queue: 64810
 
 gcn:
   server: gcn.nasa.gov
