+++ services/tns_queue/supervisor.conf
@@ -1,0 +1,5 @@
+[program:tns_queue]
+command=/usr/bin/env python services/tns_queue/tns_queue.py %(ENV_FLAGS)s
+environment=PYTHONPATH=".",PYTHONUNBUFFERED="1"
+stdout_logfile=log/tns.log
+redirect_stderr=true
