+++ skyportal/handlers/api/sources_confirmed_in_gcn.py
@@ -6,9 +6,10 @@ from tornado.ioloop import IOLoop
 from baselayer.app.access import auth_or_token, permissions
 from ..base import BaseHandler
 from .source import get_sources, MAX_SOURCES_PER_PAGE
-from .tns import tns_submission
+
 from ...models import GcnEvent, Localization, SourcesConfirmedInGCN, TNSRobot
 from ...utils.UTCTZnaiveDateTime import UTCTZnaiveDateTime
+from ...utils.tns import post_tns
 
 
 class Validator(Schema):
@@ -684,14 +685,16 @@ class SourcesConfirmedInGCNTNSHandler(BaseHandler):
                     loop = asyncio.new_event_loop()
                     asyncio.set_event_loop(loop)
 
+                request_body = {
+                    'obj_ids': [obj.obj_id for obj in sources_in_gcn],
+                    'tnsrobot_id': tnsrobot.id,
+                    'user_id': self.associated_user_object.id,
+                    'reporters': reporters,
+                }
+
                 IOLoop.current().run_in_executor(
                     None,
-                    lambda: tns_submission(
-                        [obj.obj_id for obj in sources_in_gcn],
-                        tnsrobot.id,
-                        self.associated_user_object.id,
-                        reporters=reporters,
-                    ),
+                    lambda: post_tns(request_body, timeout=30),
                 )
                 return self.success()
